Refactor/ 출력 메세지들 전체를 `GameView.java` 파일에서 클래스로 관리
