--- java-oop-3/src/view/ErrorMessage.java
@@ -1,7 +1,0 @@
-package view;
-
-public class ErrorMessage {
-    public void errorMessage() {
-        System.out.println("없는 번호입니다.");
-    }
-}
