--- java-oop-3/src/view/ExitMessage.java
@@ -1,7 +1,0 @@
-package view;
-
-public class ExitMessage {
-    public void exitMessage() {
-        System.out.println("프로그램을 종료합니다.");
-    }
-}
