--- java-oop-3/src/view/PrintMenu.java
@@ -1,11 +1,0 @@
-package view;
-
-public class PrintMenu {
-    public void printMenu() {
-        System.out.println("캐릭터를 선택하세요.");
-        System.out.println("1. Warrior");
-        System.out.println("2. Mage");
-        System.out.println("3. 종료하기");
-        System.out.print("선택: ");
-    }
-}
